Fix slim cmd to include extra proxy cmds
